Files for part 14, https://www.youtube.com/watch?v=HS1Kp1WJWBU

+++ src/drivers/mouse.cpp
@@ -81,7 +81,7 @@ void printf(char*);
         {
             if(buffer[1] != 0 || buffer[2] != 0)
             {
-                handler->OnMouseMove(buffer[1], -buffer[2]);
+                handler->OnMouseMove((int8_t)buffer[1], -((int8_t)buffer[2]));
             }
 
             for(uint8_t i = 0; i < 3; i++)
